Вынос логики на уровень выше
